requirements.txt
python-dotenv
spotipy
openai
edge-tts
pyttsx3
requests
pydub
simpleaudio